WIP: Relocating components, services, etc into their appropriate modules.
