+++ src/app/app-routing.module.ts
@@ -1,19 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
-import { RegisterUserComponent } from './register-user/register-user.component';
-
 
 const routes: Routes = [
-  {
-    path: 'login',
-    component: LoginComponent
-  }, 
-  {
-    path: 'register',
-    component: RegisterUserComponent
-  }, 
   {
     path: 'admin',
     loadChildren: () => import('./modules/admin/admin.module').then(mod => mod.AdminModule)
