+++ src/app/app.module.ts
@@ -16,28 +16,25 @@ import { AdminModule } from './modules/admin/admin.module';
 // Angular Material Modules
 import { MatButtonModule,
   MatCardModule,
-  MatFormFieldModule, 
+  MatFormFieldModule,
   MatIconModule,
-  MatInputModule, 
-  MatListModule, 
-  MatMenuModule, 
-  MatSidenavModule, 
+  MatInputModule,
+  MatListModule,
+  MatMenuModule,
+  MatSidenavModule,
   MatSnackBarModule,
   MatToolbarModule
 } from '@angular/material';
 import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
-import { RegisterUserComponent } from './register-user/register-user.component';
 import { AuthModule } from './modules/auth/auth.module';
 
 /**
  * This method is used to allow case insensitive paths to be used in url paths
- * @class LowerCaseUrlSerializer
  */
 export class LowerCaseUrlSerializer extends DefaultUrlSerializer {
   parse(url: string): UrlTree {
     return super.parse(url.toLowerCase());
-  };
+  }
 }
 
 // JWT Config
@@ -52,9 +49,7 @@ const jwtConfig = {
 @NgModule({
   declarations: [
     AppComponent,
-    HomeComponent,
-    LoginComponent,
-    RegisterUserComponent
+    HomeComponent
   ],
   imports: [
     BrowserModule,
@@ -70,12 +65,12 @@ const jwtConfig = {
     // Material
     MatButtonModule,
     MatCardModule,
-    MatFormFieldModule, 
+    MatFormFieldModule,
     MatIconModule,
-    MatInputModule, 
-    MatListModule, 
-    MatMenuModule, 
-    MatSidenavModule, 
+    MatInputModule,
+    MatListModule,
+    MatMenuModule,
+    MatSidenavModule,
     MatSnackBarModule,
     MatToolbarModule
   ],
