+++ src/app/modules/admin/admin-routing.module.ts
@@ -5,7 +5,6 @@ import { AdminComponent } from './admin/admin.component';
 import { AuthGuard } from '../auth/auth.guard';
 import { UsersComponent } from './users/users.component';
 
-
 const routes: Routes = [
   {
     path: '',
