+++ src/app/modules/auth/interfaces/ilogin-form-data.ts
@@ -1,0 +1,4 @@
+export interface ILoginFormData {
+    email: string;
+    password: string;
+}
