+++ src/app/modules/auth/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './ilogin-form-data';
+export * from './iregistration-form-data';
+export * from './itoken-user';
