+++ src/app/modules/auth/interfaces/iregistration-form-data.ts
@@ -1,0 +1,7 @@
+export interface IRegistrationFormData {
+    firstName: string;
+    lastName: string;
+    email: string;
+    password: string;
+    passwordConfirm: string;
+}
