+++ src/app/modules/auth/login/login.component.ts
@@ -2,10 +2,10 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { AuthService } from '../services/auth.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
-import { CustomValidatorService } from '../services/custom-validator.service';
-import { FormValidationService } from '../services/form-validation.service';
+import { CustomValidatorService } from '../../../services/custom-validator.service';
+import { FormValidationService } from '../../../services/form-validation.service';
 import { Subscription } from 'rxjs';
-import { ILoginFormData } from '../interfaces/auth';
+import { ILoginFormData } from '../interfaces/';
 import { MatSnackBar, MatSnackBarRef, SimpleSnackBar } from '@angular/material';
 
 @Component({
@@ -21,19 +21,18 @@ export class LoginComponent implements OnInit, OnDestroy {
   public formErrors = {
     email: '',
     password: ''
-  }
-
+  };
 
   constructor(private authService: AuthService,
-    private fb: FormBuilder,
-    private activatedRoute: ActivatedRoute,
-    private customValidator: CustomValidatorService,
-    private formValidation: FormValidationService,
-    private router: Router,
-    private snackBar: MatSnackBar) { }
+              private fb: FormBuilder,
+              private activatedRoute: ActivatedRoute,
+              private customValidator: CustomValidatorService,
+              private formValidation: FormValidationService,
+              private router: Router,
+              private snackBar: MatSnackBar) { }
 
   ngOnInit() {
-    this.returnUrl = this.activatedRoute.snapshot.queryParams['returnUrl'] || '/';
+    this.returnUrl = this.activatedRoute.snapshot.queryParams.returnUrl || '/';
     this.loginForm = this.fb.group({
       email: ['', [Validators.required, Validators.email]],
       password: ['', [Validators.required, this.customValidator.passwordValidator()]]
@@ -45,13 +44,13 @@ export class LoginComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy(): void {
-    for(let sub of this.subscriptions) {
+    for (const sub of this.subscriptions) {
       sub.unsubscribe();
     }
   }
 
   submit() {
-    if (this.snackRef !== undefined) this.snackRef.dismiss();
+    if (this.snackRef !== undefined) { this.snackRef.dismiss(); }
     this.formValidation.markFormGroupTouched(this.loginForm);
 
     if (!this.loginForm.valid) {
@@ -60,12 +59,12 @@ export class LoginComponent implements OnInit, OnDestroy {
       const formData: ILoginFormData = {
         email: this.loginForm.get('email').value,
         password: this.loginForm.get('password').value
-      }
+      };
 
       this.subscriptions.push(this.authService.login(formData)
         .subscribe(
-          res => { this.router.navigate([this.returnUrl])},
-          err => { this.snackRef = this.snackBar.open(err, 'Close')}
+          res => { this.router.navigate([this.returnUrl]); },
+          err => { this.snackRef = this.snackBar.open(err, 'Close'); }
         ));
     }
   }
