+++ src/app/modules/auth/register/register.component.html
@@ -1,1 +1,48 @@
-<p>register works!</p>
+<div id="reg-container">
+    <ng-container *ngIf="!registrationConfirmed">
+      <form [formGroup]="registrationForm" (ngSubmit)="submit()">
+        <mat-form-field>
+          <input matInput placeholder="First Name" formControlName="firstName">
+          <mat-error *ngIf="formErrors.firstName">{{ formErrors.firstName }}</mat-error>
+        </mat-form-field>
+      
+        <mat-form-field>
+          <input matInput placeholder="Last Name" formControlName="lastName">
+          <mat-error *ngIf="formErrors.lastName">{{ formErrors.lastName }}</mat-error>
+        </mat-form-field>
+      
+        <mat-form-field>
+          <input matInput placeholder="Email" formControlName="email" type="email">
+          <mat-error *ngIf="formErrors.email">{{ formErrors.email }}</mat-error>
+        </mat-form-field>
+      
+        <mat-form-field>
+          <input matInput placeholder="Password" formControlName="password" type="password">
+          <mat-error *ngIf="formErrors.password">{{ formErrors.password }}</mat-error>
+        </mat-form-field>
+      
+        <mat-form-field>
+          <input matInput placeholder="Confirm Password" formControlName="passwordConfirm" type="password">
+          <mat-hint align="end">Must match password</mat-hint>
+          <mat-error *ngIf="formErrors.passwordConfirm">{{ formErrors.passwordConfirm }}</mat-error>
+        </mat-form-field>
+      
+        <div class="button-row">
+          <button mat-raised-button (click)="cancel()" type="button">Cancel</button>
+          <button mat-raised-button color="primary" type="submit">Register</button>
+        </div>
+      </form>
+    </ng-container>
+  
+    <ng-container *ngIf="registrationConfirmed">
+      <mat-card class="course-card mat-elevation-z10">
+        <mat-card-header>
+          <mat-card-title>Registration Complete</mat-card-title>
+        </mat-card-header>
+        
+        <mat-card-content>
+          Your account registration has been submitted.  You will be sent an email confirmation once your account has been activated.
+        </mat-card-content>
+      </mat-card>
+    </ng-container>
+  </div>
