+++ src/app/modules/auth/register/register.component.scss
@@ -1,0 +1,16 @@
+#reg-container {
+    width: 400px;
+    margin: 0 auto;
+}
+
+.mat-form-field {
+    display: block;
+}
+
+.course-card {
+    margin: 20px 10px;
+}
+
+.course-actions {
+    text-align: center;
+}
