--- src/app/register-user/register-user.component.scss
@@ -1,16 +1,0 @@
-#reg-container {
-    width: 400px;
-    margin: 0 auto;
-}
-
-.mat-form-field {
-    display: block;
-}
-
-.course-card {
-    margin: 20px 10px;
-}
-
-.course-actions {
-    text-align: center;
-}
